requirements.txt
flask
python-dotenv
requests
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
pypdf
